Added the finalize_iterations() method for unitflow_iterator to get rigth pressure
In unitflow_iteration of unitflow_iterator for linear case we update the
Uh field using param*normalrhs_Uh but not the Ph. Normally because we
are solving velocity/pressure together we have to update Ph as well. But
it doesn't affect the solution here. (Because the important thing is
augmented lagrangian rhs (T-aG) which depends only on new Uh and not Ph.
But we have to use right Ph for saving pressure to file.